Add discard_url_scheme, discard_url_host, discard_url_port, discard_url_path option
If set to `true`, scheme (use `url_scheme` key) will be added to the record. Defaults to `false`.

If set to `true`, host (use `url_host` key) will be added to the record. Defaults to `false`.

If set to `true`, post (use `url_port` key) will be added to the record. Defaults to `false`.

If set to `true`, path (use `url_path` key) will be added to the record. Defaults to `false`.